# GameDebugClient/src/api/game.swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Rest Adventure - Game API",
    "version": "0.0.0-alpha.0.233+2a1139010fada7f6864f85e7bd96490563a587b1"
  },
  "servers": [
    {
      "url": "https://localhost:7056"
    }
  ],
  "paths": {
...
    "/game/team/characters/{characterGuid}/pve/monsters": {
      "get": {
        "tags": [
          "Pve"
        ],
        "summary": "Get monsters",
        "operationId": "Pve_GetMonsters",
        "parameters": [
          {
            "name": "characterGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MonsterGroup"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
    "/game/team/characters/{characterGuid}/pve/monsters/{groupId}": {
      "post": {
        "tags": [
          "Pve"
        ],
        "summary": "Attack monsters",
        "operationId": "Pve_AttackMonsters",
        "parameters": [
          {
            "name": "characterGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
    "/game/team/characters/{characterGuid}/pve/{monsterGroupGuid}/join": {
      "post": {
        "tags": [
          "Pve"
        ],
        "summary": "Join combat",
        "operationId": "Pve_JoinCombat",
        "parameters": [
          {
            "name": "characterGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "monsterGroupGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "MonsterGroup": {
        "allOf": [
          {
            "$ref": "#/components/schemas/MonsterGroupMinimal"
          },
          {
            "type": "object",
            "description": "Monster group\n            ",
            "additionalProperties": false,
            "required": [
              "attacked",
              "canAttackOrJoin"
            ],
            "properties": {
              "attacked": {
                "type": "boolean",
                "description": "Is the monster group being attacked\n            "
              },
              "canAttackOrJoin": {
                "type": "boolean",
                "description": "Can the character attack the monsters\n            "
              },
              "whyCannotAttackOrJoin": {
                "type": "string",
                "description": "Why cannot the character attack the monsters\n            ",
                "nullable": true
              }
            }
          }
        ]
      },
      "MonsterGroupMinimal": {
        "type": "object",
        "description": "Monster group (minimal)\n            ",
        "additionalProperties": false,
        "required": [
          "id",
          "monsters",
          "expectedExperience"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique ID of the group\n            ",
            "format": "guid",
            "minLength": 1
          },
          "monsters": {
            "type": "array",
            "description": "The monsters in the group\n            ",
            "items": {
              "$ref": "#/components/schemas/MonsterInGroup"
            }
          },
          "expectedExperience": {
            "type": "integer",
            "description": "The expected experience gain if the character defeats the monster group\n            ",
            "format": "int32"
          }
        }
      },
      "MonsterInGroup": {
        "type": "object",
        "description": "Monster in group\n            ",
        "additionalProperties": false,
        "required": [
          "species",
          "level"
        ],
        "properties": {
          "species": {
            "description": "The species of the monster\n            ",
            "oneOf": [
              {
                "$ref": "#/components/schemas/MonsterSpecies"
              }
            ]
          },
          "level": {
            "type": "integer",
            "description": "The level of the monster\n            ",
            "format": "int32"
          }
        }
      },
      "MonsterSpecies": {
        "type": "object",
        "description": "Monster species\n            ",
        "additionalProperties": false,
        "required": [
          "family",
          "name"
        ],
        "properties": {
          "family": {
            "description": "The family of monsters\n            ",
            "oneOf": [
              {
                "$ref": "#/components/schemas/MonsterFamily"
              }
            ]
          },
          "name": {
            "type": "string",
            "description": "The name of the species\n            ",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "description": "The description of the species\n            ",
            "nullable": true
          }
        }
      },
      "MonsterFamily": {
        "type": "object",
        "description": "Monster family\n            ",
        "additionalProperties": false,
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the family\n            ",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "description": "The description of the family\n            ",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
...
}